Senior Project 1/backend/requirements.txt
# Core FastAPI and server
fastapi>=0.100.0
uvicorn>=0.23.0

# Supabase client
supabase>=2.0.0

# Authentication and security
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# Data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# HTTP client
httpx>=0.24.0

# Environment configuration
python-dotenv>=1.0.0

# Date handling
python-dateutil>=2.8.0